requirements.txt
asgiref
Django
django-core
django-views
django-urls
pytz
sqlparse
python-multipart
pymysql
cryptography
django_auto_logout
requests_html
scikit-learn
SciPy
matplotlib
pandas
seldon-core
django_prometheus


